Move url and srcset functionality to their own modules
